# backend/docs/swagger.yaml
components:
  schemas:
    handler.coverUploadURLBody:
      properties:
        content_type:
          type: string
        file_name:
          type: string
      type: object
    handler.createEventBody:
      properties:
        name:
          type: string
        slug:
          type: string
      type: object
...
    handler.updateEventBody:
      properties:
        cover_image_key:
          type: string
        description:
          type: string
        ends_at:
          type: string
        name:
          type: string
        slug:
          type: string
        starts_at:
          type: string
        timezone:
          type: string
        venue_address:
          type: string
        venue_name:
          type: string
      type: object
...
    internal_catalog_openapi.MessageData:
      properties:
        message:
          type: string
      type: object
...
    openapi.EnvelopeCoverUploadURL:
      properties:
        data:
          $ref: '#/components/schemas/storage.CoverUploadResult'
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
    openapi.EnvelopeEventDetail:
      properties:
        data:
          $ref: '#/components/schemas/service.EventDetail'
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
    openapi.EnvelopeEventList:
      properties:
        data:
          items:
            $ref: '#/components/schemas/service.EventListItem'
          type: array
          uniqueItems: false
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
...
    openapi.EnvelopeMessage:
      properties:
        data:
          $ref: '#/components/schemas/internal_catalog_openapi.MessageData'
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
...
    platform.APIError:
      properties:
        code:
          type: string
        details: {}
        message:
          type: string
      type: object
    platform.Envelope:
      properties:
        data: {}
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
...
    service.EventListItem:
      properties:
        created_at:
          type: string
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        starts_at:
          type: string
        status:
          type: string
      type: object
...
    storage.CoverUploadResult:
      properties:
        object_key:
          type: string
        public_url:
          type: string
        upload_url:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: Staff session token as Bearer <session_id>. The Staff BFF also
        accepts an httpOnly cookie.
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: |-
    Ticket POS backend API.
    JSON endpoints use the standard envelope with data, error, and request_id.
    /health is the only endpoint that returns a bare JSON object.
  title: Ticket POS API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/staff/events:
    get:
      description: Lists catalog events for the active organization.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeEventList'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
      security:
      - BearerAuth: []
      summary: List events
      tags:
      - staff
    post:
      description: Creates a draft catalog event.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.createEventBody'
                description: Event details
                summary: body
        description: Event details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeEventDetail'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Conflict
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - staff
  /api/v1/staff/events/{id}:
    delete:
      description: Deletes a draft catalog event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeMessage'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Conflict
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - staff
    get:
      description: Returns a catalog event by ID.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeEventDetail'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get event
      tags:
      - staff
    patch:
      description: Updates catalog event fields.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.updateEventBody'
                description: Event fields
                summary: body
        description: Event fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeEventDetail'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Conflict
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - staff
  /api/v1/staff/events/{id}/cancel:
    post:
      description: Cancels a published catalog event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeEventDetail'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Conflict
      security:
      - BearerAuth: []
      summary: Cancel event
      tags:
      - staff
  /api/v1/staff/events/{id}/cover-upload-url:
    post:
      description: Returns a presigned PUT URL for uploading an event cover image.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.coverUploadURLBody'
                description: Upload details
                summary: body
        description: Upload details
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeCoverUploadURL'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Create cover upload URL
      tags:
      - staff
  /api/v1/staff/events/{id}/publish:
    post:
      description: Publishes a draft catalog event when required fields and ticket
        types are present.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeEventDetail'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Conflict
      security:
      - BearerAuth: []
      summary: Publish event
      tags:
      - staff
...